+++ crates/tauri-plugin-drag/build.rs
@@ -1,0 +1,5 @@
+const COMMANDS: &[&str] = &["start_drag"];
+
+fn main() {
+    tauri_plugin::Builder::new(COMMANDS).build();
+}
