+++ crates/tauri-plugin-drag/permissions/default.toml
@@ -1,0 +1,5 @@
+"$schema" = "schemas/schema.json"
+
+[default]
+description = "Default permissions for the plugin"
+permissions = ["allow-start-drag"]
