+++ examples/tauri/capabilities/default.json
@@ -1,0 +1,7 @@
+{
+  "$schema": "../gen/schemas/desktop-schema.json",
+  "identifier": "default",
+  "description": "Capability for the all window",
+  "windows": ["*"],
+  "permissions": ["core:default"]
+}
