+++ examples/tauri/src/lib.rs
@@ -1,0 +1,6 @@
+fn main() {
+    tauri::Builder::default()
+        .plugin(tauri_plugin_drag::init())
+        .run(tauri::generate_context!("./tauri.conf.json"))
+        .expect("failed to run app");
+}
