+++ examples/tauri/src/main.rs
@@ -1,6 +1,6 @@
+// Prevents additional console window on Windows in release, DO NOT REMOVE!!
+#![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
+
 fn main() {
-    tauri::Builder::default()
-        .plugin(tauri_plugin_drag::init())
-        .run(tauri::generate_context!("./tauri.conf.json"))
-        .expect("failed to run app");
+    tauri_app_lib::run()
 }
