+++ examples/winit/Cargo.toml
@@ -5,5 +5,5 @@ publish = false
 edition = "2021"
 
 [target."cfg(not(target_os = \"linux\"))".dependencies]
-drag = { path = "../../crates/drag" }
-winit = { version = "0.29", default-features = false, features = [ "rwh_05" ] }
+drag.workspace = true
+winit = { workspace = true, default-features = false, features = [ "rwh_05" ] }
