+++ examples/wry-dragout/Cargo.toml
@@ -5,9 +5,9 @@ publish = false
 edition = "2021"
 
 [dependencies]
-base64 = "0.21.6"
-drag = { path = "../../crates/drag" }
-serde = "1.0.195"
-serde_json = "1.0.111"
-wry = "0.24"
-dunce = "1"
+base64.workspace = true
+drag.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+wry.workspace = true
+dunce.workspace = true
